Clarify VVVVVV stack memory layout is 2.2 and below
It's only in VVVVVV 2.2 and below that critical information is put on
the stack, and has memory addresses that are not fixed. In 2.3, all
global classes got moved off of the stack and onto the static memory
region (not the heap), so everything now has a fixed memory address that
can be trivially obtained directly from the binary. (This also got rid
of having to pass all the global classes around in arguments, and fixed
a potential stack overflow on Windows due to Windows's small stack size.
Moving everything off the stack is a really good thing all around.)

So, clarify this in the README, so people who look at VVVVVV for
themselves after reading the README don't get confused about the memory
layout (since it changed in 2.3).